MSYS packages for grep and its dependency pcre
The build requirements for grep are a recent c++ compiler
and cvs (for automatic update of autotools files).

The reason why this needs to be msys grep instead of a
mingw grep is that the latter fails for
echo "abcd/" | grep -E "/$"
which is most likely due to posix path conversion as
echo "abcd/" | grep -E ".*/$"
works.

And exactly that usage of grep is present in the mgwport scripts
which makes a mingw grep unsuitable for development aka everything.

Signed-off-by: Thomas Braun <thomas.braun@virtuell-zuhause.de>