/etc/profile: set HOME to the user's home- or profile-directory
The MSYS profile script tries to create "/home/$LOGNAME" if HOME is not
set. However, on Windows, a normal user cannot create files or directories
in the installation directory.

Set the HOME variable to $HOMEDRIVE$HOMEPATH if it exists and is a
directory, otherwise to $USERPROFILE. This is how current Git for Windows
shells, TortoiseGit and GitExtensions do it as well.

Picked from msysgit, but dropped the warning line about msysGit issue 108.
The old Goole issue tracker has been deleted, and the issue is fixed by
msys-core patch "msys.dll: basic Unicode support".

Signed-off-by: Karsten Blees <blees@dcon.de>